+++ pyasic/miners/device/models/goldshell/Byte/__init__.py
@@ -13,4 +13,4 @@
 #  See the License for the specific language governing permissions and         -
 #  limitations under the License.                                              -
 # ------------------------------------------------------------------------------
-from .Byte import Byte
+from .byte import Byte
