+++ pyasic/miners/device/models/goldshell/mini_doge/__init__.py
@@ -13,4 +13,4 @@
 #  See the License for the specific language governing permissions and         -
 #  limitations under the License.                                              -
 # ------------------------------------------------------------------------------
-from .MiniDoge import GoldshellMiniDoge
+from .mini_doge import MiniDoge
