+++ pyasic/miners/device/models/goldshell/mini_doge/mini_doge.py
