+++ pyasic/miners/goldshell/bfgminer/Byte/__init__.py
@@ -13,4 +13,4 @@
 #  See the License for the specific language governing permissions and         -
 #  limitations under the License.                                              -
 # ------------------------------------------------------------------------------
-from .Byte import GoldshellByte
+from .byte import GoldshellByte
