+++ pyasic/miners/goldshell/bfgminer/mini_doge/__init__.py
@@ -13,4 +13,4 @@
 #  See the License for the specific language governing permissions and         -
 #  limitations under the License.                                              -
 # ------------------------------------------------------------------------------
-from .MiniDoge import MiniDoge
+from .mini_doge import GoldshellMiniDoge
