+++ pyasic/miners/goldshell/bfgminer/mini_doge/mini_doge.py
